recalculate tetronimo offsets
The previous offset data assumed coordinate system has a y-axis that points
"down" as the y-value ascends.

During a refactor, the coordinate space was modified so that the y-axis instead
points "up" as the y-value ascends.

The old tetronimo offsets were written/generated with the assumption that the
former system was in use.